🐳 chore: 라우팅 경로 재수정 - (기존) product/[type]/[category]/[_id]/detail - (수정) product/[type] 형태로 개선 - (수정) search/result/[_id]/detail 구조로 개선
