--- src/app/(exploring)/product/[type]/[category]/page.tsx
@@ -1,11 +1,0 @@
-export default function ProductListByCategory() {
-  return (
-    <>
-      <h1>2차 카테고리 상품 목록 페이지</h1>
-      <h2>veggie : 채소</h2>
-      <h2>fruit : 과일</h2>
-      <h2>grain : 쌀/곡류</h2>
-      <h2>mushroom : 버섯</h2>
-    </>
-  );
-}
