+++ src/app/(exploring)/search/result/[_id]/detail/page.tsx
