+++ src/app/page.tsx
@@ -13,6 +13,7 @@ export default async function Home() {
       <Navigation />
       <SlideBanner />
       <ImageCategory />
+      <ProductLinkItem link="/product/crop/veggie" linkTitle="전체 농산물" subTxt="보러 가기" />
       <main className="flex flex-col">
         <Title title="온 가족이 함께 즐기는 텃밭 체험" content="7월 한 달 간 체험 상품 20% 할인!" type="experience" />
         <ProductLinkItem
@@ -32,7 +33,7 @@ export default async function Home() {
         {/* <Title title="톡! 터지는 옥수수의 반전 매력" content="7월 제철 채소 옥수수🌽" type="gardening" /> */}
       </main>
       <footer>
-        <Link href="/product/crop/veggie/1/detail">상세페이지로 이동하기 (임시)</Link>
+        <Link href="/search/result/1/detail">상세페이지로 이동하기 (임시)</Link>
       </footer>
       {/* <Image
         src="/images/image.png"
